requirements.txt
flask
python-dotenv
gotrue
supabase
